django_project/requirements.txt
appdirs==1.4.4
apt-clone==0.2.1
apturl==0.5.2
asgiref==3.8.1
async-timeout==4.0.1
attrs==21.2.0
Babel==2.8.0
beautifulsoup4==4.10.0
beniget==0.4.1
blinker==1.4
Bottleneck==1.3.2
Brlapi==0.8.3
Brotli==1.0.9
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
distro==1.7.0
Django==5.0.6
django-ckeditor==6.7.1
django-js-asset==2.2.0
et-xmlfile==1.0.1
eyeD3==0.8.10
filelock==3.6.0
fonttools==4.29.1
fs==2.4.12
gast==0.5.2
html5lib==1.1
httplib2==0.20.2
idna==3.3
ifaddr==0.1.7
IMDbPY==2021.4.18
importlib-metadata==4.6.4
iniconfig==1.1.1
jdcal==1.0
jeepney==0.7.1
Jinja2==3.0.3
keyring==23.5.0
kiwisolver==1.3.2
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
llvmlite==0.38.0
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
matplotlib==3.5.1
more-itertools==8.10.0
mpmath==0.0.0
mutagen==1.45.1
nemo-emblems==6.0.1
netaddr==0.8.0
netifaces==0.11.0
numba==0.55.1
numexpr==2.8.1
numpy==1.21.5
oauthlib==3.2.0
odfpy==1.4.2
onboard==1.4.1
openpyxl==3.0.9
packaging==21.3
PAM==0.4.2
pandas==1.3.5
pexpect==4.8.0
Pillow==9.0.1
pluggy==0.13.0
ply==3.11
protobuf==3.12.4
psutil==5.9.0
ptyprocess==0.7.0
py==1.10.0
pycairo==1.20.1
pycryptodomex==3.11.0
pycups==2.0.1
pycurl==7.44.1
pyelftools==0.27
Pygments==2.11.2
PyGObject==3.42.1
PyICU==2.8.1
pyinotify==0.9.6
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyparted==3.11.7
pyRFC3339==1.1
pytest==6.2.5
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-gnupg==0.4.8
python-magic==0.4.24
python-xlib==0.29
pythran==0.10.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
qrcode==7.3.1
reportlab==3.6.8
requests==2.25.1
requests-file==1.5.1
scipy==1.8.0
SecretStorage==3.3.1
setproctitle==1.2.2
six==1.16.0
soupsieve==2.3.1
sqlparse==0.5.0
sympy==1.9
systemd-python==234
tables==3.7.0
tinycss2==1.1.1
tldextract==3.1.2
toml==0.10.2
typing_extensions==4.12.0
ubuntu-drivers-common==0.0.0
ufoLib2==0.13.1
ufw==0.36.1
unicodedata2==14.0.0
Unidecode==1.3.3
urllib3==1.26.5
wadllib==1.3.6
webencodings==0.5.1
websockets==9.1
xdg==5
xkit==0.0.0
xlrd==1.2.0
xlwt==1.3.0
yt-dlp==2022.4.8
zipp==1.0.0
